requirements.txt
Flask
Gunicorn
requests
python-dotenv
flask-restx